ICC compiles for Xeon Phi.

+++ simulation/grid.h
@@ -14,7 +14,7 @@ template<size_t cellCapacity> struct List : mref<uint16> {
 };
 
 struct Grid {
- static constexpr size_t cellCapacity = 32;
+ static constexpr int cellCapacity = 32;
  const vec3 scale;
  const vec3 min, max;
  const int3 size = ::max(int3(5,5,1), int3(::floor(scale*(max-min)))+int3(2));
